requirements.txt
dlt[duckdb]>=0.5.0
pandas>=2.0.0
pyarrow>=15.0.0
python-dotenv>=1.0.0